requirements.txt
akshare>=1.18.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
faiss-cpu>=1.7.0